+++ packages/react/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ponder/react
 
+## 0.10.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.10.5
+
+## 0.10.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.10.4
+
+## 0.10.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.10.3
+
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.10.2
+
 ## 0.10.1
 
 ### Patch Changes